Consolidate resources#create under purls#create
