+++ app/controllers/application_controller.rb
@@ -12,8 +12,8 @@ class ApplicationController < ActionController::Base
   end
 
   def druid_param
-    druid = params.require(:druid)
-    druid = "druid:#{druid}" unless druid.start_with?('druid:')
+    druid = params[:druid]
+    druid = "druid:#{druid}" unless druid&.start_with?('druid:')
     druid
   end
 end
